GH-20: Run the tests for both versions of antd
