+++ tests/antd4.test.tsx
@@ -1,0 +1,8 @@
+import Form from "antd4/lib/form";
+import Button from "antd4/lib/button";
+import FormItem from "antd4/lib/form/FormItem";
+
+import commonTests from "./common";
+import PhoneInput from "../src/index4";
+
+commonTests(PhoneInput, Form, FormItem, Button);
