+++ tests/antd5.test.tsx
@@ -1,0 +1,8 @@
+import Form from "antd/lib/form";
+import Button from "antd/lib/button";
+import FormItem from "antd/lib/form/FormItem";
+
+import commonTests from "./common";
+import PhoneInput from "../src/index5";
+
+commonTests(PhoneInput, Form, FormItem, Button);
